requirements.txt
python-dotenv==1.0.0
streamlit==1.36.0
google-generativeai==0.7.0
Flask==3.0.3
Flask-Cors==4.0.1
firebase-admin==6.5.0
PyJWT==2.8.0
Werkzeug==3.0.0
requests==2.31.0
